update releaser script

+++ scripts/releaser.sh
@@ -1,20 +1,29 @@
-#!/bin/bash -x
+#!/bin/bash
+shopt -s extglob
 
-version="${1//[[:space:]]/}"
+version="$(jq -r .version package.json)"
 
 if [ -z "$version" ]; then
-    echo "Must specify a release version"
+    echo "package.json does not contain a version" >&2
     exit 1
 fi
 
-git stash
-git fetch origin --tags
-jq --arg version "$version" '.version = $version' package.json >package.json.tmp
-mv package.json.tmp package.json
-yarn build
-git add lib
-git commit -m "release files for version $version"
-git tag "v$version"
-git rm -rf lib
-git reset --soft HEAD~1
-git stash pop
+version="${version##+(v)}" # discard leading v
+
+(
+    # Runs in subshell so that the git operations do not affect the rest of the script
+    git fetch origin --tags
+    if [ "$(git tag -l "v$version")" ]; then
+        echo "Version $version already exists" >&2
+        exit 1
+    fi
+    echo "Releasing version $version" >&2
+    git stash
+    yarn build
+    git add lib
+    git commit -m "release files for version $version"
+    git tag "v$version"
+    git rm -rf lib
+    git reset --soft HEAD~1
+    git stash pop
+)
